@@ -28,76 +28,76 @@
         "./packages/source"
     ],
     "devDependencies": {
-        "@babel/cli": "^7.16.0",
-        "@babel/core": "^7.9.0",
-        "@babel/plugin-proposal-class-properties": "^7.8.3",
-        "@babel/preset-env": "^7.16.4",
-        "@babel/preset-react": "^7.8.3",
-        "@babel/preset-typescript": "^7.9.0",
+        "@babel/cli": "^7.27.2",
+        "@babel/core": "^7.27.4",
+        "@babel/plugin-proposal-class-properties": "^7.18.6",
+        "@babel/preset-env": "^7.27.2",
+        "@babel/preset-react": "^7.27.1",
+        "@babel/preset-typescript": "^7.27.1",
         "@faker-js/faker": "^7.6.0",
-        "@linaria/babel-preset": "^4.5.3",
-        "@linaria/cli": "^4.5.3",
-        "@linaria/core": "^4.5.3",
-        "@linaria/esbuild": "^4.5.3",
-        "@linaria/react": "^4.5.3",
+        "@linaria/babel-preset": "^4.5.4",
+        "@linaria/cli": "^4.5.4",
+        "@linaria/core": "^4.5.4",
+        "@linaria/esbuild": "^4.5.4",
+        "@linaria/react": "^4.5.4",
         "@linaria/shaker": "^4.5.3",
-        "@linaria/vite": "^4.5.3",
-        "@linaria/webpack-loader": "^4.5.3",
-        "@storybook/addon-docs": "^9.0.1",
-        "@storybook/addon-links": "^9.0.1",
-        "@storybook/addons": "^7.6.17",
-        "@storybook/react-vite": "^9.0.1",
-        "@storybook/react-webpack5": "^9.0.1",
+        "@linaria/vite": "^4.5.4",
+        "@linaria/webpack-loader": "^4.5.4",
+        "@storybook/addon-docs": "^9.0.11",
+        "@storybook/addon-links": "^9.0.11",
+        "@storybook/addons": "^7.6.20",
+        "@storybook/react-vite": "^9.0.11",
+        "@storybook/react-webpack5": "^9.0.11",
         "@testing-library/dom": "^10.4.0",
         "@testing-library/react": "^16.3.0",
         "@testing-library/react-hooks": "^8.0.1",
-        "@testing-library/user-event": "^14.5.1",
-        "@types/cheerio": "^0.22.30",
-        "@types/lodash": "^4.14.177",
-        "@types/marked": "^4.0.2",
-        "@types/node": "^18.7.13",
-        "@types/prosemirror-commands": "^1.0.4",
-        "@types/prosemirror-inputrules": "^1.0.4",
-        "@types/prosemirror-model": "^1.16.1",
-        "@types/prosemirror-state": "^1.2.8",
+        "@testing-library/user-event": "^14.6.1",
+        "@types/cheerio": "^0.22.35",
+        "@types/lodash": "^4.17.17",
+        "@types/marked": "^4.3.2",
+        "@types/node": "^18.19.112",
+        "@types/prosemirror-commands": "^1.3.0",
+        "@types/prosemirror-inputrules": "^1.2.0",
+        "@types/prosemirror-model": "^1.17.0",
+        "@types/prosemirror-state": "^1.4.0",
         "@types/react": "^16.14.21 || 17.x || 18.x || 19.x",
         "@types/react-dom": "^16.9.5 || 17.x || 18.x || 19.x",
-        "@types/react-syntax-highlighter": "^15.5.5",
-        "@typescript-eslint/eslint-plugin": "^5.30.5",
-        "@typescript-eslint/parser": "^6.4.1",
-        "@typescript-eslint/typescript-estree": "^5.30.5",
-        "@vitejs/plugin-react": "^4.2.0",
-        "babel-loader": "^8.1.0",
+        "@types/react-syntax-highlighter": "^15.5.13",
+        "@typescript-eslint/eslint-plugin": "^5.62.0",
+        "@typescript-eslint/parser": "^6.21.0",
+        "@typescript-eslint/typescript-estree": "^5.62.0",
+        "@vitejs/plugin-react": "^4.5.2",
+        "babel-loader": "^8.4.1",
         "chokidar-cli": "^3.0.0",
-        "clean-css-cli": "^5.6.2",
+        "clean-css-cli": "^5.6.3",
         "concurrently": "^8.2.2",
-        "esbuild": "^0.19.2",
-        "eslint": "^8.19.0",
-        "eslint-plugin-import": "^2.26.0",
-        "eslint-plugin-react": "^7.30.1",
-        "eslint-plugin-react-hooks": "^4.6.0",
-        "eslint-plugin-sonarjs": "^0.13.0",
-        "eslint-plugin-unicorn": "^43.0.1",
+        "esbuild": "^0.25.5",
+        "eslint": "^8.57.1",
+        "eslint-plugin-import": "^2.31.0",
+        "eslint-plugin-react": "^7.37.5",
+        "eslint-plugin-react-hooks": "^4.6.2",
+        "eslint-plugin-sonarjs": "^0.25.1",
+        "eslint-plugin-unicorn": "^43.0.2",
         "http-server": "^14.1.1",
         "jsdom": "^22.1.0",
         "lodash": "^4.17.21",
-        "marked": "^4.0.10",
-        "mini-css-extract-plugin": "^2.6.1",
-        "prettier": "^3.0.2",
+        "marked": "^4.3.0",
+        "mini-css-extract-plugin": "^2.9.2",
+        "prettier": "^3.5.3",
         "react": "^19.1.0",
         "react-dom": "^19.1.0",
-        "react-laag": "^2.0.3",
-        "react-responsive-carousel": "^3.2.7",
-        "react-syntax-highlighter": "^15.4.5",
+        "react-laag": "^2.0.5",
+        "react-responsive-carousel": "^3.2.23",
+        "react-syntax-highlighter": "^15.6.1",
         "react-test-renderer": "^19.1.0",
         "resolve-typescript-plugin": "^2.0.1",
-        "storybook": "^9.0.1",
-        "tsc-esm-fix": "^2.7.8",
-        "typedoc": "^0.25.4",
-        "typescript": "^5.1.6",
+        "storybook": "^9.0.11",
+        "tsc-esm-fix": "^2.20.27",
+        "typedoc": "^0.28.5",
+        "typescript": "^5.8.3",
         "vitest": "^0.34.6",
         "vitest-canvas-mock": "^0.3.3",
-        "webpack": "^5.73.0"
+        "webpack": "^5.99.9"
     },
     "overrides": {
         "storybook": "$storybook"
