@@ -51,25 +51,24 @@
     ],
     "dependencies": {
         "@glideapps/glide-data-grid": "6.0.4-alpha9",
-        "@linaria/react": "^4.5.3",
-        "@toast-ui/editor": "3.1.10",
-        "@toast-ui/react-editor": "3.1.10",
-        "react-select": "^5.8.0"
+        "@linaria/react": "^4.5.4",
+        "@toast-ui/editor": "3.2.2",
+        "@toast-ui/react-editor": "3.2.3",
+        "react-select": "^5.10.1"
     },
     "peerDependencies": {
         "react": "^16.12.0 || 17.x || 18.x || 19.x",
         "react-dom": "^16.12.0 || 17.x || 18.x || 19.x"
     },
     "devDependencies": {
-        "@babel/cli": "^7.16.0",
-        "@types/prosemirror-commands": "^1.0.4",
-        "@types/react": "16.14.21",
-        "eslint": "^8.19.0",
-        "eslint-plugin-import": "^2.22.0",
-        "eslint-plugin-react": "^7.21.5",
-        "eslint-plugin-react-hooks": "^4.2.0",
+        "@babel/cli": "^7.27.2",
+        "@types/prosemirror-commands": "^1.3.0",
+        "eslint": "^8.57.1",
+        "eslint-plugin-import": "^2.31.0",
+        "eslint-plugin-react": "^7.37.5",
+        "eslint-plugin-react-hooks": "^4.6.2",
         "react-resize-detector": "^7.1.2",
-        "tsc-esm-fix": "^2.7.8",
-        "typescript": "^5.1.6"
+        "tsc-esm-fix": "^2.20.27",
+        "typescript": "^5.8.3"
     }
 }
