@@ -57,9 +57,9 @@
     },
     "homepage": "https://grid.glideapps.com",
     "dependencies": {
-        "@linaria/react": "^4.5.3",
+        "@linaria/react": "^4.5.4",
         "canvas-hypertxt": "^1.0.3",
-        "react-number-format": "^5.0.0"
+        "react-number-format": "^5.4.4"
     },
     "peerDependencies": {
         "lodash": "^4.17.19",
@@ -69,13 +69,13 @@
         "react-responsive-carousel": "^3.2.7"
     },
     "devDependencies": {
-        "@babel/cli": "^7.16.0",
+        "@babel/cli": "^7.27.2",
         "@glideapps/ts-helper": "5.0.0",
         "@skypack/package-check": "^0.2.2",
         "@vitest/coverage-v8": "^0.34.6",
-        "eslint": "^8.19.0",
+        "eslint": "^8.57.1",
         "react-resize-detector": "^7.1.2",
-        "tsc-esm-fix": "^2.7.8",
-        "typedoc": "^0.23.13"
+        "tsc-esm-fix": "^2.20.27",
+        "typedoc": "^0.28.5"
     }
 }
